Move createStore out of legacy
