+++ packages/satcheljs/lib/index.ts
@@ -15,7 +15,7 @@ export { default as Middleware } from './legacy/Middleware';
 export { default as ActionFunction } from './legacy/ActionFunction';
 export { default as ActionContext } from './legacy/ActionContext';
 export { default as DispatchFunction } from './legacy/DispatchFunction';
-export { default as createStore } from './legacy/createStore';
+export { default as createStore } from './createStore';
 export { default as action } from './legacy/action';
 export { default as select, SelectorFunction } from './legacy/select';
 export { default as createUndo, UndoResult, CreateUndoReturnValue} from './legacy/createUndo';
